+++ src/main/java/com/openclassrooms/ApiRentals/config/WebConfig.java
@@ -1,5 +1,8 @@
 package com.openclassrooms.ApiRentals.config;
 
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
@@ -10,6 +13,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class WebConfig {
 
+	Logger logger = LoggerFactory.getLogger(WebConfig.class);
 	
     @Bean
     public WebMvcConfigurer corsConfigurer()  {
@@ -26,9 +30,10 @@ public class WebConfig {
             
             @Override
             public void addResourceHandlers(ResourceHandlerRegistry registry) {
-                // Configure the resource handler to serve files from the "uploads" directory
-            	String uploadsPath = "file:///C:/Users/mbenziane/Repo Git/python/Java/ApiRentals/src/main/resources/static/uploads/";
-                //logger.info("Serving static files from: {}", uploadsPath);
+                // Configurer le gestionnaire de ressources pour qu'il serve les fichiers du répertoire « uploads ».
+            	//chemin qui point vers le dossier static du projet Spring.
+            	String uploadsPath ="classpath:/static/uploads/";
+                
                 registry.addResourceHandler("/api/uploads/**")
                         .addResourceLocations(uploadsPath);
             }
