+++ src/main/java/com/openclassrooms/ApiRentals/service/UserService.java
@@ -7,8 +7,6 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import com.openclassrooms.ApiRentals.dto.UserDTO;
 import com.openclassrooms.ApiRentals.model.User;
 import com.openclassrooms.ApiRentals.repository.UsersRepository;
 
@@ -38,17 +36,6 @@ public class UserService {
     }    
    
     
-    public UserDTO convertToDTO(User user) {
-        return new UserDTO(
-            user.getId(),
-            user.getUsername(),
-            user.getEmail(),
-            user.getPassword(),
-            user.getCreated_at(),
-            user.getUpdated_at()
-        );
-    }
-    
     public User findByUsername(String name) {
     	
         return usersRepository.findUserByName(name);
@@ -57,7 +44,6 @@ public class UserService {
     public Optional<User> findById(Long id) {
         return usersRepository.findById(id);
     }
-
 
 
 	public User findByEmail(String email) {
